Moved Button components into own folder and updated files that used Button
