+++ src/features/goals/Goal.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faPen } from '@fortawesome/free-solid-svg-icons';
-import Button from '../../components/Button';
+import Button from '../../components/Button/Button';
 import './Goals.css';
 
 export default function Goal({ goal, removeGoal, completeGoal, toggleEdit, updateGoal }) {
