+++ src/features/settings/Settings.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCog } from '@fortawesome/free-solid-svg-icons';
-import Button from '../../components/Button';
+import Button from '../../components/Button/Button';
 import { 
   selectMilitaryTime,
   toggleTime,
